add dt_SQL_subtract to DateMethods1
This commit adds the new method dt_SQL_subtract to
DBIx::Class::Helper::ResultSet::DateMethods1

It works exactly like dt_SQL_add, but subtracts dates (surprise!).

While one can pass negative values to dt_SQL_add to subtract a literal
value ($amount), this does not work when the value is not passed as a
literal value but as the name of another column where the actual value
is stored. Hence this new method.

- dt_SQL_subtract tests for Postgres, MySQL and SQLite
- add sponsor note